backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
pandas
scikit-learn
